1_graph_creation/functions/agent3_extract_rules/requirements.txt
functions-framework==3.*
flask==2.*
google-genai
